Move kebab items to kebab menu folder
